fixed delete button

+++ holdings-tab.go
@@ -13,6 +13,7 @@ import (
 
 // holdingsTab returns a container containing the holdings table.
 func (app *Config) holdingsTab() *fyne.Container {
+	app.Holdings = app.getHoldingSlice()
 	app.HoldingsTable = app.getHoldingsTable()
 
 	holdingsContainer := container.NewBorder(
@@ -30,26 +31,25 @@ func (app *Config) holdingsTab() *fyne.Container {
 // It formats the holdings data into a slice of slices of interface{} using getHoldingSlice().
 // Each row in the table corresponds to a holding, with the last cell containing a delete button.
 func (app *Config) getHoldingsTable() *widget.Table {
-	data := app.getHoldingSlice()
-	app.Holdings = data
-
 	table := widget.NewTable(
 		func() (int, int) {
-			return len(data), len(data[0])
+			return len(app.Holdings), len(app.Holdings[0])
 		},
 		func() fyne.CanvasObject {
 			ctr := container.NewVBox(widget.NewLabel(""))
 			return ctr
 		},
 		func(i widget.TableCellID, o fyne.CanvasObject) {
-			if i.Col == (len(data[0])-1) && i.Row != 0 {
+			if i.Col == (len(app.Holdings[0])-1) && i.Row != 0 {
 				// last cell - put in a button
 				w := widget.NewButtonWithIcon("Delete", theme.DeleteIcon(), func() {
 					dialog.ShowConfirm("Delete?", "", func(deleted bool) {
-						id, _ := strconv.Atoi(data[i.Row][0].(string))
-						err := app.DB.DeleteHolding(int64(id))
-						if err != nil {
-							app.ErrorLog.Println(err)
+						if deleted {
+							id, _ := strconv.Atoi(app.Holdings[i.Row][0].(string))
+							err := app.DB.DeleteHolding(int64(id))
+							if err != nil {
+								app.ErrorLog.Println(err)
+							}
 						}
 						app.refreshHoldingsTable()
 					}, app.MainWindow)
@@ -62,7 +62,7 @@ func (app *Config) getHoldingsTable() *widget.Table {
 			} else {
 				// text info
 				o.(*fyne.Container).Objects = []fyne.CanvasObject{
-					widget.NewLabel(data[i.Row][i.Col].(string)),
+					widget.NewLabel(app.Holdings[i.Row][i.Col].(string)),
 				}
 			}
 		})
